🚚(frontend) move styled box to disable events
Move the styled box to disable events in the appropriate styled components
folder.
